feature(split): add support for sensors from peripheral
This commit adds a new GATT characteristics on the peripheral side
and wires it up to read sensor values. The central side subscribes
to this new characteristics and replays sensor values on its side.

—

This commit was originally made by Stephen Wan. I just adjusted it so that it rebases on top of later changes on the zmk main branch.