implement gpio_165 driver
This is an input-only gpio driver for use with 74x165
style parallel-input shift registers, (using SPI peripheral)

The user can choose to either use cs-gpios of spi or
load-gpios for connecting to the SH/LD pin of the shift
register. The difference is that cs-gpios will keep
the input latch enabled while idle, and load-gpios will
instead pulse the SH/LD signal before the register
content is shifted. In both cases the gpio should be
configured GPIO_ACTIVE_HIGH.