refactor(encoder): use behavior queue to archieve key taps for encoder
Remove k_msleep() call used in sensor trigger callback that's meant to simulate a key
tap (press and release), instead we can use the existing behavior queue
mechanism, where a press and a release events are queued up separately.
The trick is to add .binging_pressed and .binding_released callbacks for
the sensor driver api, which will gets called when the queued events are
processed accordingly.