feat(ble): Add option to disable profile auto save
Add Kconfig option to disable persisting the BLE profile index after
every change. This is useful if the user wants to have the same profile
connected at startup regardless of which profile was connected last.

Add `&bt BT_SAVE` command to manually persist the currently active BLE
profile, ensuring it is selected as the active profile on startup.