feat(display): Always display current BLE profile
Display USB or BLE active transport symbol (USB if BLE is disabled)

To the right, If BLE is enabled, always show the current BLE profile
index and its connection status (not bonded / disconnected / connected).

The previous behavior was to show the USB symbol only, if that was the
active transport. It's useful to see the active BLE profile index and
status even if it is not the active transport. It is what you will
connect to if you toggle USB to BLE. It is possible to browse for a free
profile without unplugging USB.