fix(core): support magic bootloader values for nRF52833
Enable ZMK_BOOTMODE_MAGIC_VALUE_BOOTLOADER_TYPE_ADAFRUIT_NRF52 for
nRF52833, add nrf52833_uf2_boot_mode.dtsi as a copy of the nRF52840 file
for easy inclusion.

Enable retention for all nrfmicro subvariants (as this is the only
in-tree board that uses the nRF52833).

This fixes &bootloader for nRF52833 boards post-4.1 upgrade.